chore(pipeline): expose `TestStage` and add test stage set
